backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7
coverage==7.6.12
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
flake8==7.1.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.33.0
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.0
mysqlclient==2.2.7
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
psutil==7.0.0
psycopg==3.2.5
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PySocks==1.7.1
pytest==8.3.5
pytest-cov==4.1.0
pytest-django==4.10.0
python-dotenv==1.0.1
pywin32==308; sys_platform == "win32"
pyzmq==26.2.1
requests==2.32.3
selenium==4.29.0
setuptools==75.8.2
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
djangorestframework-simplejwt==5.5.0
pillow==11.1.0